Rollback to float32

+++ layers/gated_gcn.py
@@ -44,7 +44,7 @@ class GatedGCN_1d(nn.Module):
         if in_channels != out_channels:
             self.residual = False
 
-        dtype=torch.float16
+        dtype=torch.float32
 
         self.A_1 = nn.Linear(in_channels, out_channels, dtype=dtype)
         self.A_2 = nn.Linear(in_channels, out_channels, dtype=dtype)
@@ -111,13 +111,13 @@ class GatedGCN_1d(nn.Module):
         g.ndata['h'] = h
         g.edata['e'] = e
 
-        g.ndata['A1h'] = self.A_1(h).type(torch.float32)
-        g.ndata['A2h'] = self.A_2(h).type(torch.float32)
-        g.ndata['A3h'] = self.A_3(h).type(torch.float32)
+        g.ndata['A1h'] = self.A_1(h) # .type(torch.float32)
+        g.ndata['A2h'] = self.A_2(h) # .type(torch.float32)
+        g.ndata['A3h'] = self.A_3(h) # .type(torch.float32)
 
-        g.ndata['B1h'] = self.B_1(h).type(torch.float32)
-        g.ndata['B2h'] = self.B_2(h).type(torch.float32)
-        g.edata['B3e'] = self.B_3(e).type(torch.float32)
+        g.ndata['B1h'] = self.B_1(h) # .type(torch.float32)
+        g.ndata['B2h'] = self.B_2(h) # .type(torch.float32)
+        g.edata['B3e'] = self.B_3(e) # .type(torch.float32)
 
         g_reverse = dgl.reverse(g, copy_ndata=True, copy_edata=True)
 
@@ -214,7 +214,7 @@ class GatedGCN_backwards(nn.Module):
         if in_channels != out_channels:
             self.residual = False
 
-        dtype = torch.float16
+        dtype = torch.float32
 
         self.A_1 = nn.Linear(in_channels, out_channels, dtype=dtype)
         self.A_2 = nn.Linear(in_channels, out_channels, dtype=dtype)
@@ -235,13 +235,13 @@ class GatedGCN_backwards(nn.Module):
         g.ndata['h'] = h
         g.edata['e'] = e
 
-        g.ndata['A1h'] = self.A_1(h).type(torch.float32)
-        g.ndata['A2h'] = self.A_2(h).type(torch.float32)
-        g.ndata['A3h'] = self.A_3(h).type(torch.float32)
+        g.ndata['A1h'] = self.A_1(h) # .type(torch.float32)
+        g.ndata['A2h'] = self.A_2(h) # .type(torch.float32)
+        g.ndata['A3h'] = self.A_3(h) # .type(torch.float32)
 
-        g.ndata['B1h'] = self.B_1(h).type(torch.float32)
-        g.ndata['B2h'] = self.B_2(h).type(torch.float32)
-        g.edata['B3e'] = self.B_3(e).type(torch.float32)
+        g.ndata['B1h'] = self.B_1(h) # .type(torch.float32)
+        g.ndata['B2h'] = self.B_2(h) # .type(torch.float32)
+        g.edata['B3e'] = self.B_3(e) # .type(torch.float32)
 
         g_reverse = dgl.reverse(g, copy_ndata=True, copy_edata=True)
